05/03/2021 - 5:30am Final Submission
Final version that is being uploaded to Blackboard.

+++ Secure Software - Assignment 1/TemperatureSensor.h
@@ -5,19 +5,19 @@ class TemperatureSensor
 {
 private:
 	std::string name = "Temperature Sensor";
-	int currentTemp = 15;
-	int minTemp = -50;
-	int maxTemp = 100;
+	int16_t currentTemp = 15;
+	int16_t minTemp = -50;
+	int16_t maxTemp = 100;
 
 public:
 	std::string getName() const;
-	int getCurrentTemp() const noexcept;
-	void setCurrentTemp(int newCurrentTemp) noexcept;
-	void configureTempSensor(int newMaxTemp, int newMinTemp) noexcept;
+	int16_t getCurrentTemp() const noexcept;
+	void setCurrentTemp(int16_t newCurrentTemp) noexcept;
+	void configureTempSensor(int16_t newMaxTemp, int16_t newMinTemp) noexcept;
 
-	int getMinTemp() const noexcept;
-	int getMaxTemp() const noexcept;
+	int16_t getMinTemp() const noexcept;
+	int16_t getMaxTemp() const noexcept;
 
-	void setMinTemp(int newMinTemp) noexcept;
-	void setMaxTemp(int newMaxTemp) noexcept;
+	void setMinTemp(int16_t newMinTemp) noexcept;
+	void setMaxTemp(int16_t newMaxTemp) noexcept;
 };
