+++ Secure Software - Assignment 1/Validation.h
@@ -9,6 +9,6 @@ public:
     uint16_t unsignedSecureAddition(uint16_t ui_a, uint16_t ui_b);
     uint16_t unsignedSecureSubtraction(uint16_t ui_a, uint16_t ui_b);
 
-    int signedSecureSubtraction(int si_a, int si_b);
-    int signedSecureAddition(int si_a, int si_b);
+    int64_t signedSecureSubtraction(int64_t si_a, int64_t si_b);
+    int64_t signedSecureAddition(int64_t si_a, int64_t si_b);
 };
